refactor(dashboard): consolidate normalize utilities, remove lab-unified wrapper
Move generic utilities (asString, asNumber, asInt, asBoolean) from
api/trpg.ts to components/common/normalize.ts as single source of truth.
Add TypeScript overloads to support both strict (no fallback) and lenient
(with fallback) call signatures without breaking existing callers.

Remove lab-unified.ts wrapper (17-line passthrough with unused LabUnified
alias) — dashboard-shell.ts now lazy-imports Lab directly.
